move pickled loading function to be a bit more PEP8ish and avoid the FileUtils directory
