+++ PYME/IO/compatibility.py
